requirements.txt
click>=8.0.4
click_shell>=2.1
prettytable>=3.2.0
PyHackTheBox>=0.5.6
pytest>=7.0.1
pytest-cov>=3.0.0
pytest-subprocess>=1.4.1                         
pytest-xdist>=2.5.0 
setuptools
twine
